cannot delete appointment if it is booked

+++ rightnow-firebase/src/context/businessContext.js
@@ -51,6 +51,8 @@ export default class BusinessProvider extends Component {
     },
 
     delete_appointment: () => {
+      if (!this.state.selected_appointment.is_available) return false;
+
       const appt_id = this.state.selected_appointment.id;
 
       firebase.firestore().collection("_appointment_").doc(appt_id).delete()
@@ -59,6 +61,8 @@ export default class BusinessProvider extends Component {
       firebase.firestore().collection("_business_").doc(this.state.uid)
         .collection("future_appointments").doc(appt_id).delete()
         .then(res => console.log("success", res)).catch(err => console.log("error", err));
+
+      return true;
     }
   }
